Use STDIN as source for purty format
Purty >4.1.0 supports STDIN as input source. This way we can use the
text from the document directly and send it to the purty process. Purty
will then respond with the formated content and we can replace the
content in the editor. This way the file does not have to be saved
before it is formated and VSCode's "format on save" should work as
expected.